pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "banana-agi"
version = "0.1.0"
description = "ARC solver using vision models"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.32.0",
    "matplotlib>=3.8.0",
    "numpy>=2.3.2",
    "pillow>=11.3.0",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "tqdm>=4.67.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/banana_agi"]

[project.scripts]
arc-solver = "banana_agi.arc_solver:main"
